+++ src/data-flow/service-registration.js
@@ -1,4 +1,5 @@
 import { ServiceCollection } from '../foundation/di/service-collection.js';
+import { CommandFlowDescriptor } from './commands/command-flow-descriptor.js';
 import { ExecutionOverviewDescriptor } from './execution/execution-overview-descriptor.js';
 
 export function addDataFlow(services) {
@@ -6,6 +7,7 @@ export function addDataFlow(services) {
     throw new Error('addDataFlow expects an instance of ServiceCollection.');
   }
 
+  services.registerSingleton('CommandFlowDescriptor', () => new CommandFlowDescriptor());
   services.registerSingleton('ExecutionOverviewDescriptor', () => new ExecutionOverviewDescriptor());
 
   return services;